// src/app/ngrx/bootstrap/data/bootstrap-component-names.enum.ts
export enum BootstrapComponentNamesEnum {
  Alert = 'alert',
  Badge = 'badge',
  Breadcrumbs = 'breadcrumb',
  Button = 'button',
  ButtonGroup = 'buttongroup',
  Card = 'card',
  Carousel = 'carousel',
  Collapse = 'collapse',
  Dropdown = 'dropdown',
  Form = 'form',
  InputGroup = 'inputgroup',
  Jumbotron = 'jumbotron',
  ListGroup = 'listgroup',
  Modal = 'modal',
  Navs = 'navs',
  Navbar = 'navbar',
  Pagination = 'pagination',
  Popovers = 'popovers',
  Progress = 'progress',
  Scrollspy = 'scrollspy',
  Tooltips = 'tooltips'
}